requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.3
Flask-WTF==1.1.1
Flask-Mail==0.9.1
Werkzeug==2.3.7
bcrypt==4.0.1
WTForms==3.0.1
Pillow==10.0.1
python-dotenv==1.0.0
email-validator==2.0.0
